Moving away from settingsModel to simplify things If layer is hidden, featureCount is 0
