+++ src/geo/map/layer-model-base.js
@@ -44,6 +44,10 @@ var MapLayer = Model.extend({
     return !!this.get('visible');
   },
 
+  isHidden: function () {
+    return !this.isVisible();
+  },
+
   toggle: function () {
     this.set('visible', !this.get('visible'));
   },
