+++ src/vis/settings.js
@@ -1,11 +1,9 @@
 var Backbone = require('backbone');
-var RenderModes = require('./render-modes');
 
 var UISettings = Backbone.Model.extend({
   defaults: {
     showLegends: true,
-    showLayerSelector: false,
-    renderMode: RenderModes.AUTO
+    showLayerSelector: false
   }
 });
