+++ test/spec/geo/leaflet/leaflet-map-view.spec.js
@@ -45,8 +45,7 @@ describe('geo/leaflet/leaflet-map-view', function () {
       el: container,
       mapModel: map,
       visModel: new Backbone.Model(),
-      layerGroupModel: this.layerGroupModel,
-      settingsModel: new Backbone.Model()
+      layerGroupModel: this.layerGroupModel
     });
 
     mapView.render();
@@ -197,8 +196,7 @@ describe('geo/leaflet/leaflet-map-view', function () {
       el: container,
       mapModel: map,
       visModel: new Backbone.Model(),
-      layerGroupModel: new Backbone.Model(),
-      settingsModel: new Backbone.Model()
+      layerGroupModel: new Backbone.Model()
     });
     mapView.render();
 
@@ -220,8 +218,7 @@ describe('geo/leaflet/leaflet-map-view', function () {
       el: container,
       mapModel: map,
       visModel: new Backbone.Model(),
-      layerGroupModel: new Backbone.Model(),
-      settingsModel: new Backbone.Model()
+      layerGroupModel: new Backbone.Model()
     });
     mapView.render();
