+++ test/spec/geo/map-view.spec.js
@@ -43,8 +43,7 @@ describe('core/geo/map-view', function () {
       visModel: new Backbone.Model(),
       layerGroupModel: new LayerGroupModel(null, {
         layersCollection: this.map.layers
-      }),
-      settingsModel: new Backbone.Model()
+      })
     });
     spyOn(this.mapView, 'setCursor');
     spyOn(this.mapView, 'getNativeMap');
@@ -66,8 +65,7 @@ describe('core/geo/map-view', function () {
         layerGroupModel: new LayerGroupModel(null, {
           windshaftMap: this.windshaftMap,
           layersCollection: this.map.layers
-        }),
-        settingsModel: new Backbone.Model()
+        })
       });
       spyOn(this.mapView, 'getNativeMap');
       spyOn(this.mapView, '_addLayerToMap');
