+++ test/spec/geo/ui/infowindow.spec.js
@@ -18,8 +18,7 @@ describe('geo/ui/infowindow-view', function() {
       mapModel: map,
       visModel: new Backbone.Model(),
       layerViewFactory: jasmine.createSpyObj('layerViewFactory', ['createLayerView']),
-      layerGroupModel: new Backbone.Model(),
-      settingsModel: new Backbone.Model()
+      layerGroupModel: new Backbone.Model()
     });
 
     model = new InfowindowModel({
@@ -293,8 +292,7 @@ describe('geo/ui/infowindow-view', function() {
         mapModel: map,
         visModel: new Backbone.Model(),
         layerViewFactory: jasmine.createSpyObj('layerViewFactory', ['createLayerView']),
-        layerGroupModel: new Backbone.Model(),
-        settingsModel: new Backbone.Model()
+        layerGroupModel: new Backbone.Model()
       });
 
       model = new InfowindowModel({
@@ -423,8 +421,7 @@ describe('geo/ui/infowindow-view', function() {
         mapModel: map,
         visModel: new Backbone.Model(),
         layerViewFactory: jasmine.createSpyObj('layerViewFactory', ['createLayerView']),
-        layerGroupModel: new Backbone.Model(),
-        settingsModel: new Backbone.Model()
+        layerGroupModel: new Backbone.Model()
       });
 
       model = new InfowindowModel({
