+++ test/spec/geo/ui/tooltip.spec.js
@@ -15,8 +15,7 @@ describe('geo/ui/tooltip-view', function () {
       mapModel: map,
       visModel: new Backbone.Model(),
       layerViewFactory: jasmine.createSpyObj('layerViewFactory', ['createLayerView']),
-      layerGroupModel: new Backbone.Model(),
-      settingsModel: new Backbone.Model()
+      layerGroupModel: new Backbone.Model()
     });
     this.mapView.render();
